WIP: Add SECP256R1 encryption
